+++ tests/org/elixir_lang/inspection/NoParentheseStrictTestCase.java
@@ -1,6 +1,9 @@
 package org.elixir_lang.inspection;
 
+import com.intellij.codeInsight.intention.IntentionAction;
 import org.elixir_lang.PlatformTestCase;
+
+import java.util.List;
 
 /**
  * Created by kadie.enheduanna.inanna on 12/6/14.
@@ -34,6 +37,32 @@ public class NoParentheseStrictTestCase extends PlatformTestCase {
         myFixture.configureByFile("QualifierDotQuoteParentheses.ex");
         myFixture.enableInspections(NoParenthesesStrict.class);
         myFixture.checkHighlighting();
+    }
+
+    public void testFunctionSpaceEmptyParenthesesQuickFix() {
+        assertQuickFixRewrites("function ()", "function()");
+    }
+
+    public void testFunctionSpacePositionalsInParenthesesQuickFix() {
+        assertQuickFixRewrites("function (one, two)", "function(one, two)");
+    }
+
+    public void testQualifierDotQuoteParenthesesQuickFix() {
+        assertQuickFixRewrites("One.\"two\" ()", "One.\"two\"()");
+    }
+
+    private void assertQuickFixRewrites(String before, String after) {
+        myFixture.configureByText("remove_space.ex", before);
+        myFixture.enableInspections(NoParenthesesStrict.class);
+
+        List<IntentionAction> quickFixes = myFixture.getAllQuickFixes();
+
+        assertEquals(1, quickFixes.size());
+        assertEquals("Remove space between function name and parentheses", quickFixes.get(0).getText());
+
+        myFixture.launchAction(quickFixes.get(0));
+
+        myFixture.checkResult(after);
     }
 
     @Override